Refactor `Optional`
