+++ README.md
@@ -1,8 +1,8 @@
-[![Maintainability Rating](https://sonarcloud.io/api/project_badges/measure?project=raccoons-co_java-base&metric=sqale_rating)](https://sonarcloud.io/summary/new_code?id=raccoons-co_java-base)
-[![codecov](https://codecov.io/gh/raccoons-co/java-base/branch/main/graph/badge.svg?token=FtCvNhCrBK)](https://codecov.io/gh/raccoons-co/java-base)
-[![CircleCI](https://dl.circleci.com/status-badge/img/gh/raccoons-co/java-base/tree/master.svg?style=svg)](https://dl.circleci.com/status-badge/redirect/gh/raccoons-co/java-base/tree/master)
+[![Maintainability Rating](https://sonarcloud.io/api/project_badges/measure?project=raccoons-co_meeko&metric=sqale_rating)](https://sonarcloud.io/summary/new_code?id=raccoons-co_meeko)
+[![codecov](https://codecov.io/gh/raccoons-co/meeko/branch/main/graph/badge.svg?token=FtCvNhCrBK)](https://codecov.io/gh/raccoons-co/meeko)
+[![CircleCI](https://dl.circleci.com/status-badge/img/gh/raccoons-co/meeko/tree/master.svg?style=svg)](https://dl.circleci.com/status-badge/redirect/gh/raccoons-co/meeko/tree/master)
 
-Java Base Refactoring
+Java Base Utils
 ---
 
 >*Always leave the code you are working on a little bit better than you found it.*
